+++ android/src/main/java/com/swmansion/reanimated/layoutReanimation/NativeMethodsHolder.java
@@ -6,6 +6,6 @@ import java.util.Map;
 public interface NativeMethodsHolder {
     public void startAnimationForTag(int tag);
     public void removeConfigForTag(int tag);
-    public Map<String, Object> getStyleWhileMounting(int tag, float progress, HashMap<String, Double> values, int depth);
-    public Map<String, Object> getStyleWhileUnmounting(int tag, float progress, HashMap<String, Double> values, int depth);
+    public Map<String, Object> getStyleWhileMounting(int tag, float progress, HashMap<String, Integer> values, int depth);
+    public Map<String, Object> getStyleWhileUnmounting(int tag, float progress, HashMap<String, Integer> values, int depth);
 }
